requirements.txt
pillow
daemon
sounddevice
wavio
soundfile
better-profanity
openai-whisper
